Refactoring

+++ algebra.c
@@ -1,12 +1,12 @@
 #include "algebra.h"
 
 /**
- *	[EN] This method check if an integer number is odd
- *	[PT] Esse metodo verifica se um numero inteiro e par
- *	@param input_number an integer number
+ *	[EN] This method check if an long long integer number is odd
+ *	[PT] Esse metodo verifica se um numero long long inteiro e par
+ *	@param input_number an long long integer number
  *	@return TRUE if is Odd or FALSE if is Even
  */
-Boolean is_odd(int input_number)
+Boolean is_odd(long long int input_number)
 {
 	Boolean is_odd = FALSE;
 	if(input_number % 2 == 1)
@@ -17,12 +17,12 @@ Boolean is_odd(int input_number)
 }
 
 /**
- *	[EN] This method check if an integer number is negativo
- *	[PT] Esse metodo verifica se um numero inteiro e negativo
- *	@param input_number an integer number
+ *	[EN] This method check if an long long integer number is negativo
+ *	[PT] Esse metodo verifica se um numero long long inteiro e negativo
+ *	@param input_number an long long integer number
  *	@return TRUE if is negative or FALSE if is natural
  */
-Boolean is_negative(int input_number)
+Boolean is_negative(long long int input_number)
 {
 	Boolean is_negative = FALSE;
 	if(input_number < 0)
@@ -38,7 +38,7 @@ Boolean is_negative(int input_number)
  *	@param 
  *	@return TRUE if is not coprime or FALSE if is coprime
  */
-Boolean is_divisible(int numerator_number, int denominator_number)
+Boolean is_divisible(long long int numerator_number, long long int denominator_number)
 {
 	Boolean is_divisible = FALSE;
 	if( (numerator_number % denominator_number) == 0)
@@ -56,10 +56,10 @@ Boolean is_divisible(int numerator_number, int denominator_number)
  *	@param maximum_number (b)
  *	@return random_number (x)
  */
-unsigned int random_natural_number(unsigned int minimum_number, unsigned int maximum_number)
+unsigned long long int random_natural_number(unsigned long long int minimum_number, unsigned long long int maximum_number)
 {
 	srand(time(NULL));
-	unsigned int random_number = (rand() % (maximum_number - minimum_number + 1)) + minimum_number;
+	unsigned long long int random_number = (rand() % (maximum_number - minimum_number + 1)) + minimum_number;
 	return random_number;
 }
 
@@ -72,16 +72,16 @@ unsigned int random_natural_number(unsigned int minimum_number, unsigned int max
  *		first_number (s)
  *		second_number (d) an odd value
  */
-Two_Natural_Numbers decompose_as_power_of_two(unsigned int input_number)
+Two_Natural_Numbers decompose_as_power_of_two(unsigned long long int input_number)
 {
-	unsigned int s = 1;
-	unsigned int d = 1;
+	unsigned long long int s = 1;
+	unsigned long long int d = 1;
 	Two_Natural_Numbers factors;
 	for(s = 1; s < input_number; s++)
 	{
 		for(d = 1; d < (input_number/s); d++)
 		{
-			int result = (int) pow(2, s)*d;
+			long long int result = (long long int) pow(2, s)*d;
 			if(result == input_number && is_odd(d) == TRUE)
 			{
 				factors.first_number = s;
@@ -98,10 +98,10 @@ Two_Natural_Numbers decompose_as_power_of_two(unsigned int input_number)
  *	@param input_number number to be calculated magnitude
  *	@return magnitude
  */
-int get_magnitude(int input_number)
+long long int get_magnitude(long long int input_number)
 {	
-	int number = input_number;
-	int magnitude = 0;
+	long long int number = input_number;
+	long long int magnitude = 0;
 	while(number > 0){
 		magnitude++;
 		number = number / 10;
